requirements.txt
nonebot2
nonebot_adapter_qq
nonebot2[fastapi]
nonebot2[httpx]
nonebot2[websockets]
apscheduler


ffmpeg-python
sqlalchemy
selenium
autopep8
graiax-silkcoder[ffmpeg]
graiax-silkcoder
fake_useragent
pycryptodome
PyExecJS
gradio_client
tortoise-orm
urllib3
zhdate
chinese_calendar
lunardate
playwright
nonebot-plugin-htmlrender
tenacity
paramiko

commonX
jmcomic
natsort

requests
pillow
pyyaml
qrcode
lazy_object_proxy
openai
typing_extensions
psutil
Beautifulsoup4
aiohttp